Tidying up merge
Move all instructions into a "whatnow" command; makes it much easier
than referring to history / script. Add some colour to the output to
more easily distinguish stuff that can be copied & pasted. Also switch
the ordering of reconstructing the changelog and updating the maintainer
to make it easier to redo the latter.