Use hardlinks or copying as required
If source and target are the same fs, then hardlink. Otherwise, just
copy regularly (should enable us to sync the books by album which is
more useful in the car). Also make threshold for by_genre configurable.

This also means that SOURCE and TARGET become paths, and everything in
the FOLDER_MAP is also a full path. In turn, this means the --source and
--target options are obsolete (as everything's absolute internally
anyway). Should probably check all paths passed in are absolute, but
it's good enough for now...