Add rebased --again
For those (frequent) cases where the tests show we need to edit the
rebase, and we wind up not just adding commits but removing or
re-working the rebased commits, add some crude "intelligence" for
rebuilding the logical/$new and candidate/$new tags based on the
generated commit subjects.

It'll probably break on anything complex, but it'll do for now.