Use afew for filtering ... for now
I'm not so enamoured of afew's filtration mechansims ... my little
mailfilter was a bit more flexible at this and nicer to configure. But
it didn't handle threads, or new flags, or some of mbsync's weirdness. I
need to rewrite it as a "proper" project and see if it can be more
useful that way ...